user-service/requirements.txt
fastapi
uvicorn[standard]
pymongo
passlib[bcrypt]
python-jose
PyJWT
python-multipart